split apart character sheet into different components
